Moved csv and ini readers into their own subfolder.
