+++ src/csv/load_csv.c
@@ -1,6 +1,6 @@
 // Generic CSV reader
 
-#include "csv.h"
+#include "load_csv.h"
 
 #include <assert.h>
 #include <stdio.h>
