+++ src/ini/ini_read.c
@@ -1,5 +1,5 @@
 
-#include "ini.h"
+#include "ini_read.h"
 
 #include <assert.h>
 #include <stddef.h>
