+++ src/load_time_averaged.h
@@ -1,7 +1,7 @@
 
 #pragma once
 
-#include "csv.h"
+#include "csv/load_csv.h"
 
 #ifndef _LOAD_TIME_AVERAGED_H_
 #  define _LOAD_TIME_AVERAGED_H_
